Move LED colours into rgb_led.py, LED_FLASH_DIAL into dial.py
rgb_led.py's COLOURS dict independently hardcoded the same 5 colour
name strings that constants.py separately defined as COLOUR_* -
duplicated vocabulary with no structural link between them. Move
COLOUR_RED/GREEN/BLUE/WHITE/OFF into rgb_led.py and build COLOURS from
them instead, so there's a single source of truth.

Also move LED_FLASH_DIAL into dial.py as _LED_FLASH_DIAL, matching
_LED_FLASH_SHORT's precedent in buttons.py - _dial_loop() is its only
caller. LED_FLASH_LONG stays in radio_config.py since it's used by two
unrelated App-level events (city latch, stream error) with no single
owning hardware module.

Cost surfaced by testing: importing _LED_FLASH_DIAL from dial.py means
main.py now transitively requires evdev importable, not just RPi.GPIO.
Extended tests/main_test.py's existing RPi.GPIO stub pattern to also
stub evdev, and documented the widened residual import gap in
ARCHITECTURE.md.
