+++ src/radioglobe/constants.py
@@ -9,10 +9,3 @@ STATUS_CALIBRATING = "Calibrating"
 STATUS_CALIBRATED = "CALIBRATED"
 STATUS_CALIBRATE = "CALIBRATE"
 STATUS_SHUTDOWN = "Shutdown"
-
-# LED colours
-COLOUR_RED = "red"
-COLOUR_GREEN = "green"
-COLOUR_BLUE = "blue"
-COLOUR_WHITE = "white"
-COLOUR_OFF = "off"
